代码提交

+++ mapping-qy108-management/mapping-qy108-service/src/main/java/com/aaa/six/base/BaseService.java
@@ -169,7 +169,7 @@ public abstract class BaseService<T> {
      *@Date: 2020/5/23 9:43
      *@Return: T
      */
-    public T queryOne(Integer id) throws Exception{
+    public T queryOne(Long id) throws Exception{
         return mapper.selectByPrimaryKey(id);
     }
